cpu/esp*: common parts of syscalls in cpu/esp_common
